+++ client/src/components/Pages/Domains/Main/Close.tsx
