+++ client/src/components/Pages/Gene/Main/Gene.scss
