+++ client/src/components/Pages/Gene/Main/Gene.tsx
@@ -1,0 +1,17 @@
+import {GeneSearch} from '../Search/GeneSearch'
+import {GeneResults} from '../Results/GeneResults'
+
+import './Gene.scss'
+
+interface Props {
+  index: number
+}
+export const Gene: React.FC<Props> = ( { index }) => {
+
+  return (
+    <div className="genetab-container">
+      <GeneSearch />
+      <GeneResults/>
+    </div>
+  )
+}
