+++ client/src/components/Pages/Gene/Search/Close.tsx
