+++ client/src/components/Pages/Gene/Search/GeneSearch.scss
@@ -1,0 +1,55 @@
+.gene-search {
+  // padding-left: 250px;
+  flex-grow: 1;
+  display:flex;
+  align-items: center;
+  // margin: 0 80px;
+  justify-content: space-around;
+  // min-height: 142px;
+  position: relative;
+}
+
+.selected-parent {
+  width: 200px;
+  height: 150px;
+
+  position: absolute;
+  top: 2%;
+  right: 20%;
+
+  ul {
+    display: flex;
+    width: 50px;
+    list-style-type: none;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    padding: 0 0 0 50px;
+    margin: 0;
+  
+    li {
+      display: flex;
+      min-width: 100%;
+      background-color: var(--primary); 
+      color: white;
+      margin: 10px 0;
+      padding: 5px 9px 5px 3px;
+      border-radius: 5%;
+      align-items: center;
+      justify-content: center;
+      height: 100%;
+  
+      span {
+        display: flex;
+        align-items: center;
+        cursor: pointer;
+  
+      }
+  
+    }
+  }
+}
+
+
+
+
