+++ client/src/components/Pages/RegElement/Main/RegElement.tsx
