+++ client/src/components/Pages/RegElement/RegElementForm/RegElementForm.scss
