+++ client/src/components/Pages/Structure/TransCompInput/TransCompInput.scss
@@ -1,0 +1,4 @@
+.MuiCard-root {
+  width: 100%;
+  background-color: pink;
+}
