+++ client/src/components/Pages/Summary/Main/Summary.scss
