+++ client/src/components/main/App/App.scss
@@ -1,8 +1,28 @@
 .App {
   height: 100vh;
   display: flex;
-  justify-content: center;
+  flex-direction: column;
   align-items: center;
+  background: linear-gradient(45deg, var(--gradient-1) 30%, var(--gradient-2)  90%);
+  margin: 0 auto;
 }
+
+.title {
+  margin: 0 auto;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  flex: 1;
+  color: var(--black-primary);
+  font-weight: normal;
+  visibility: hidden;
+}
+
+.main-component {
+  flex-grow: 1;
+}
+
+
