+++ client/src/components/main/App/App.tsx
@@ -1,5 +1,5 @@
 import React, {useState, useEffect} from 'react';
-import NavTabs from '../NavTabs/NavTabs';
+import NavTabs from '../Nav/NavTabs';
 import { SuggestionContext } from '../../../global/contexts/SuggestionContext'
 import { FusionContext } from '../../../global/contexts/FusionContext'
 import '../../../global/styles/global.scss'
@@ -26,7 +26,11 @@ function App() {
               style={{
                ...colorTheme
                } as React.CSSProperties}>
-              <NavTabs />
+              <h1 className="title">Fusion Curation</h1>
+              <div className="main-component">
+                <NavTabs />
+              </div>
+              
             </div>
           </FusionContext.Provider>
