+++ client/src/components/main/Nav/NavTabs.scss
@@ -1,0 +1,30 @@
+.nav-tabs {
+  background-color: var(--white);
+}
+
+.tab-panel {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+.sub-tab {
+  width: 80vw;
+  height: 600px;
+}
+
+.previous {
+
+}
+
+.continue{ 
+  margin-left: auto;
+}
+
+.footer {
+  flex-shrink: 0;
+  padding: 15px 10px;
+  display: flex;
+  align-items: center;
+  background-color: var(--white);
+}
