--- client/src/components/tabs/Domains/DomainForm/DomainForm.scss
@@ -1,76 +1,0 @@
-.domain-tab-container {
-  display: flex;
-  width: 100%;
-  height: 100%;
-  align-items: stretch;
-  flex: 1;
-
-  .right {
-    display: flex;
-    flex-direction: column;
-    width: 50%;
-    align-items: center;
-    justify-content: center;
-    }
-
-  .left {
-    width: 50%;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    background-color: #F5F5FA;
-  }
-
-  .blurb {
-    width: 70%;
-    font-size: xx-large;
-    font-weight: 200;
-  }
-
-  .bold {
-    font-weight: bold
-  }
-
-  .sub-blurb{
-    margin-top: 20px;
-    width: 70%;
-    font-size:larger;
-    font-weight: 200;
-  }
-
-  .close-button-reg {
-    margin-left: 38px;
-    margin-top: -3px;
-    position: absolute;
-  }
-
-  .formInput {
-    margin: 40px;
-  }
-
-  .regel {
-    margin-top: 20px;
-    width: 34%;
-    font-size:medium;
-    font-weight: 200;
-    border: 1px solid rgba(0, 0, 0, 0.3);
-    padding: 10px;
-    margin-left: -35px;
-    border-radius: 7px;
-  }
-  .blurb-container {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    padding-bottom: 40px;
-  }
-
-  .add-button {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    width: 690px;
-  }
-
-
-}
