--- client/src/components/tabs/Gene/GeneResults/GeneResults.scss
@@ -1,24 +1,0 @@
-
-.gene-results {
-  background-color: #F5F5FA;
-  min-height: 400px;
-  // display: flex;
-  // padding: 15px 0 50px 0;
-
-  h1 {
-    display: flex;
-    justify-content: center;
-  }
-
-  h4 {
-    display: flex;
-    justify-content: center;
-    font-weight: 300;
-    font-size:large;
-  }
-}
-
-
-
-
-
