--- client/src/components/tabs/Gene/GeneSearch/GeneSearch.scss
@@ -1,50 +1,0 @@
-.gene-search {
-  // padding-left: 250px;
-  flex: 1;
-  display:flex;
-  align-items: center;
-  // margin: 0 80px;
-  // justify-content: space-around;
-  // min-height: 142px;
-}
-
-.selected-parent {
-  width: 200px;
-  height: 150px;
-
-  ul {
-    display: flex;
-    width: 50px;
-    list-style-type: none;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    padding: 0 0 0 50px;
-    margin: 0;
-  
-    li {
-      display: flex;
-      min-width: 100%;
-      background-color: #5DD3D3; 
-      color: white;
-      margin: 10px 0;
-      padding: 5px 9px 5px 3px;
-      border-radius: 5%;
-      align-items: center;
-      justify-content: center;
-      height: 100%;
-  
-      span {
-        display: flex;
-        align-items: center;
-        cursor: pointer;
-  
-      }
-  
-    }
-  }
-}
-
-
-
-
