--- client/src/components/tabs/Gene/Main/Gene.tsx
@@ -1,17 +1,0 @@
-import {GeneSearch} from '../GeneSearch/GeneSearch'
-import {GeneResults} from '../GeneResults/GeneResults'
-
-import './Gene.scss'
-
-interface Props {
-  index: number
-}
-export const Gene: React.FC<Props> = ( { index }) => {
-
-  return (
-    <div className="genetab-container">
-      <GeneSearch />
-      <GeneResults/>
-    </div>
-  )
-}
