--- client/src/components/tabs/Gene/StructureDiagram/StructureDiagram.scss
@@ -1,68 +1,0 @@
-
-
-
-.structure {
-  padding: 5px;
-  display: inline-block;
-  background-color: white;
-  display: flex;
-  justify-content: space-around;
-  width: 250px;
-  margin: 7px;
-  cursor: pointer;
-  border: 5px solid white;
-
-  .gene  {
-    background-color: #79FFFF;
-    color: #79FFFF;
-    width: 120px;
-  }
-  .transcript_segment {
-    background-color: #fbfcbb;
-    color: #fbfcbb;
-    width: 80px;
-  }
-  .linker_sequence {
-    background-color: #FF94FF;
-    color: #FF94FF;
-    width: 25px;
-  }
-  .genomic_region {
-    background-color: #CCCCFF;
-    color: #CCCCFF;
-    width: 25px;
-  }
-
-  &:hover {
-    border: 5px solid rgba(127, 127, 127, .15);
-    
-    .gene {
-      background-color: #bbfafa;
-      color: #bbfafa;
-      width: 120px;
-    }
-    .transcript_segment {
-      background-color: #fcfcc0;
-      color: #fcfcc0;
-      width: 80px;
-    }
-    .linker_sequence {
-      background-color: #ffbaff;
-      color: #ffbaff;
-      width: 25px;
-    }
-    .genomic_region {
-      background-color: #e1e1fc;
-      color: #e1e1fc;
-      width: 25px;
-    }
-  }
-}
-
-.highlighted {
-  border: 5px solid #0BD3D3;
-  &:hover {
-    border: 5px solid #0BD3D3;
-  }
-}
-
