--- client/src/components/tabs/Structure/Builder/Builder.scss
@@ -1,123 +1,0 @@
-.builder {
-  display: flex;
-  align-items: stretch;
-  background: #F5F5FA;
-  flex: 1;
-
-  .gene {
-    background:#79FFFF;
-  }
-  
-  .transcript_segment {
-    background-color: #FEFF91;
-  }
-  
-  .linker_sequence {
-    background-color: #FF94FF;
-  }
-  
-  .genomic_region {
-    background-color: #CCCCFF;
-  }
-
-
-.block-container {
-  text-align: center;
-  flex: 1;
-  overflow: scroll;
-}
-
-.block {
-  display: flex;
-  text-align: center;
-  justify-content: center;
-  font-weight: 300;
-  padding: 10px;
-  margin: 0 50px 15px 50px;
-	font-size: 18px;
-	border-radius: 5px;
-}
-
-h3 {
-  text-align: center;
-  font-weight: 500;
-}
-
-.options {
-  background-color: darkgray;
-  width: 30%;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.option-item {
-  display: flex;
-  text-align: center;
-  justify-content: center;
-  font-weight: 300;
-  padding: 10px;
-  margin: 0 50px 15px 50px;
-	font-size: 18px;
-	border-radius: 5px;
-
-  .gene {
-    border: 2px solid #bbfafa !important;
-    color: #757575;
-    font-weight: 500
-  }
-  
-  .transcript_component{
-    border: 2px solid #fcfcc0 !important;
-    color: #757575;
-    font-weight: 500
-  }
-  
-  .linker_sequence {
-    border: 2px solid #fcbaf1 !important;
-    color: #757575;
-    font-weight: 500
-  }
-  
-  .genomic_region {
-    border: 2px solid #e1e1fc !important;
-    color: #757575;
-    font-weight: 500
-  }
-}
-
-
-
-
-.card-parent{
-  display: flex;
-  flex-direction: row;
-  justify-content: space-around;
-  width: 100%;
-}
-
-.input-parent {
-  display: flex;
-  flex-direction: column;
-  width: 300px;
-}
-
-.top-inputs {
-  display: flex;
-  justify-content: space-between;
-}
-
-.bottom-inputs {
-  display: flex;
-  justify-content: space-between;
-}
-
-.buttons {
-  width: 40px;
-  margin-left: 20px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-  
-}
