--- client/src/components/tabs/Structure/Main/Structure.scss
@@ -1,67 +1,0 @@
-.structure-tab-container {
-  display: flex;
-  flex-direction: column;
-  align-items: stretch;
-  height: 100%;
-
-  .structure-summary {
-    .gene {
-      background-color: #79FFFF;
-      width: 320px;
-      text-align: center;
-      color: black;
-    }
-    
-    .transcript_segment {
-      background-color: #FEFF91;
-      width: 150px;
-      text-align: center;
-      color: black;
-    }
-    .linker_sequence {
-      background-color: #FF94FF;
-      width: 60px;
-      text-align: center;
-    }
-    
-    .genomic_region {
-      background-color: #CCCCFF;
-      width: 60px;
-      text-align: center;
-    }
-    
-  }
-
-
-}
-
-// .summary-title {
-//   h3 {
-//     font-weight: 400;
-//   } 
-// }
-
-
-
-
-// .structure-summary {
-//   padding: 5px;
-//   display: inline-block;
-//   background-color: white;
-//   display: flex;
-//   justify-content: space-around;
-//   width: 600px;
-//   margin: 5px;
-//   cursor: pointer;
-// }
-
-// .summary-container {
-//   display: flex;
-//   justify-content: center;
-//   min-width: 100%;
-//   padding-bottom: 15px;
-// }
-
-// .structure-container {
-//   min-width: 100%;
-// }
