+++ client/src/global/styles/global.scss
@@ -3,9 +3,9 @@
 body {
   font-family: 'Poppins', sans-serif;
   min-height: 100vh;
+  margin: 0 auto;
 }
 
 html {
-  background: linear-gradient(45deg, #FE6B8B 30%, #FF8E53 90%);
-  min-height: 100vh;;
+  min-height: 100vh;
 }
