+++ client/testAPI/data.json
@@ -22,6 +22,7 @@
         "component_type": "transcript_segment",
         "component_name": "refseq:NM_152263.3",
         "component_id": "7c4684d2-5830-45f9-8e31-8278ae59bd50",
+        "shorthand": "NM_152263.3",
         "exon_start": 1,
         "exon_start_offset": 0,
         "exon_end": 8,
@@ -38,6 +39,7 @@
         "component_type": "genomic_region",
         "component_id": "95bd881d-3bb4-46c5-a3f5-d07c6a77e46e",
         "id": "refseq:NM_152263.3_exon1-exon8",
+        "shorthand": "e1e8",
         "type": "LocationDescriptor",
         "location": {
           "sequence_id": "ga4gh:SQ.ijXOSP3XSsuLWZhXQ7_TJ5JXu4RJO6VT",
@@ -58,6 +60,7 @@
       {
         "component_type": "gene",
         "component_name": "ALK",
+        "shorthand": "ALK",
         "component_id": "2ad09f78-832e-4158-a13c-cd020365b7f6",
         "gene_descriptor": {
           "id": "gene:ALK",
